Move address validation to validation module
