+++ includes/validation/merge-fields.php
@@ -8,6 +8,7 @@
 namespace Mailchimp\WordPress\Includes\Validation;
 
 use WP_Error;
+use stdClass;
 
 /**
  * Validate phone
@@ -73,4 +74,42 @@ function merge_validate_phone( $opt_val, $data ) {
 	}
 
 	return $opt_val;
+}
+
+/**
+ * Validate address
+ *
+ * @param array $opt_val Option value
+ * @param array $data Data
+ * @return mixed
+ */
+function mailchimp_sf_merge_validate_address( $opt_val, $data ) {
+	if ( 'Y' === $data['required'] ) {
+		if ( empty( $opt_val['addr1'] ) || empty( $opt_val['city'] ) ) {
+			/* translators: %s: field name */
+			$message = sprintf( esc_html__( 'You must fill in %s.', 'mailchimp' ), esc_html( $data['name'] ) );
+			$error   = new WP_Error( 'invalid_address_merge', $message );
+			return $error;
+		}
+		/**
+		 * Condition
+		 * 1) If the address is not required and
+		 * 2) address line 1 or city is empty
+		 *
+		 * Result
+		 * Then don't send address to Mailchimp
+		 * Return false skips merge formatting in `mailchimp_sf_merge_submit`
+		 */
+	} elseif ( empty( $opt_val['addr1'] ) || empty( $opt_val['city'] ) ) {
+		return false;
+	}
+
+	$merge          = new stdClass();
+	$merge->addr1   = $opt_val['addr1'];
+	$merge->addr2   = $opt_val['addr2'];
+	$merge->city    = $opt_val['city'];
+	$merge->state   = $opt_val['state'];
+	$merge->zip     = $opt_val['zip'];
+	$merge->country = $opt_val['country'];
+	return $merge;
 }
